+++ scripts/DBUpdate-to-3.3.mssql.sql
@@ -1,5 +1,5 @@
 -- ----------------------------------------------------------
---  driver: mssql, generated: 2013-04-04 11:16:16
+--  driver: mssql, generated: 2013-04-29 14:48:37
 -- ----------------------------------------------------------
 -- ----------------------------------------------------------
 --  insert into table ticket_history_type
@@ -7,3 +7,9 @@
 INSERT INTO ticket_history_type (name, valid_id, create_by, create_time, change_by, change_time)
     VALUES
     ('TitleUpdate', 1, 1, current_timestamp, 1, current_timestamp);
+DROP INDEX dynamic_field_value.index_search_date;
+CREATE INDEX dynamic_field_value_search_date ON dynamic_field_value (field_id, value_date);
+DROP INDEX dynamic_field_value.index_search_int;
+CREATE INDEX dynamic_field_value_search_int ON dynamic_field_value (field_id, value_int);
+DROP INDEX dynamic_field_value.index_field_values;
+CREATE INDEX dynamic_field_value_field_values ON dynamic_field_value (object_id);