+++ scripts/DBUpdate-to-3.3.mysql.sql
@@ -1,5 +1,5 @@
 # ----------------------------------------------------------
-#  driver: mysql, generated: 2013-04-04 11:16:16
+#  driver: mysql, generated: 2013-04-29 14:48:37
 # ----------------------------------------------------------
 # ----------------------------------------------------------
 #  insert into table ticket_history_type
@@ -7,3 +7,9 @@
 INSERT INTO ticket_history_type (id, name, valid_id, create_by, create_time, change_by, change_time)
     VALUES
     (50, 'TitleUpdate', 1, 1, current_timestamp, 1, current_timestamp);
+DROP INDEX index_search_date ON dynamic_field_value;
+CREATE INDEX dynamic_field_value_search_date ON dynamic_field_value (field_id, value_date);
+DROP INDEX index_search_int ON dynamic_field_value;
+CREATE INDEX dynamic_field_value_search_int ON dynamic_field_value (field_id, value_int);
+DROP INDEX index_field_values ON dynamic_field_value;
+CREATE INDEX dynamic_field_value_field_values ON dynamic_field_value (object_id);