+++ scripts/DBUpdate-to-3.3.postgresql.sql
@@ -1,5 +1,5 @@
 -- ----------------------------------------------------------
---  driver: postgresql, generated: 2013-04-04 11:16:16
+--  driver: postgresql, generated: 2013-04-29 14:48:37
 -- ----------------------------------------------------------
 SET standard_conforming_strings TO ON;
 -- ----------------------------------------------------------
@@ -8,4 +8,10 @@ SET standard_conforming_strings TO ON;
 INSERT INTO ticket_history_type (name, valid_id, create_by, create_time, change_by, change_time)
     VALUES
     ('TitleUpdate', 1, 1, current_timestamp, 1, current_timestamp);
+DROP INDEX index_search_date;
+CREATE INDEX dynamic_field_value_search_date ON dynamic_field_value (field_id, value_date);
+DROP INDEX index_search_int;
+CREATE INDEX dynamic_field_value_search_int ON dynamic_field_value (field_id, value_int);
+DROP INDEX index_field_values;
+CREATE INDEX dynamic_field_value_field_values ON dynamic_field_value (object_id);
 SET standard_conforming_strings TO ON;