+++ scripts/database/otrs-schema-post.db2.sql
@@ -1,5 +1,5 @@
 -- ----------------------------------------------------------
---  driver: db2, generated: 2012-11-07 17:21:15
+--  driver: db2, generated: 2013-04-29 14:53:46
 -- ----------------------------------------------------------
 ALTER TABLE valid ADD CONSTRAINT FK_valid_create_by_id FOREIGN KEY (create_by) REFERENCES users (id);