+++ scripts/database/otrs-schema-post.oracle.sql
@@ -1,5 +1,5 @@
 -- ----------------------------------------------------------
---  driver: oracle, generated: 2012-11-07 17:21:16
+--  driver: oracle, generated: 2013-04-29 14:53:47
 -- ----------------------------------------------------------
 SET DEFINE OFF;
 ALTER TABLE valid ADD CONSTRAINT FK_valid_create_by_id FOREIGN KEY (create_by) REFERENCES users (id);