+++ scripts/database/otrs-schema-post.postgresql.sql
@@ -1,5 +1,5 @@
 -- ----------------------------------------------------------
---  driver: postgresql, generated: 2012-11-07 17:21:16
+--  driver: postgresql, generated: 2013-04-29 14:53:47
 -- ----------------------------------------------------------
 SET standard_conforming_strings TO ON;
 ALTER TABLE valid ADD CONSTRAINT FK_valid_create_by_id FOREIGN KEY (create_by) REFERENCES users (id);