+++ scripts/database/otrs-schema.oracle.sql
@@ -1,5 +1,5 @@
 -- ----------------------------------------------------------
---  driver: oracle, generated: 2012-11-07 17:21:16
+--  driver: oracle, generated: 2013-04-29 14:53:47
 -- ----------------------------------------------------------
 SET DEFINE OFF;
 -- ----------------------------------------------------------
@@ -2058,9 +2058,9 @@ end;
 /
 --;
 CREATE INDEX FK_dynamic_field_value_field90 ON dynamic_field_value (field_id);
-CREATE INDEX index_field_values ON dynamic_field_value (field_id, object_id);
-CREATE INDEX index_search_date ON dynamic_field_value (field_id, value_date);
-CREATE INDEX index_search_int ON dynamic_field_value (field_id, value_int);
+CREATE INDEX dynamic_field_value_field_va6e ON dynamic_field_value (object_id);
+CREATE INDEX dynamic_field_value_search_db3 ON dynamic_field_value (field_id, value_date);
+CREATE INDEX dynamic_field_value_search_int ON dynamic_field_value (field_id, value_int);
 -- ----------------------------------------------------------
 --  create table dynamic_field
 -- ----------------------------------------------------------